Moved base array to long type

+++ commands.c
@@ -3,23 +3,23 @@
 #include "engine.h"
 #include "util.h"
 
-int execute_commands_size;
-int alloc_executor_size;
-int get_executor_size;
-int set_executor_size;
-int print_executor_size;
+long execute_commands_size;
+long alloc_executor_size;
+long get_executor_size;
+long set_executor_size;
+long print_executor_size;
 void init_command_sizes() {
     /* It does not matter if these sizes are not exact. 
      * They need to be at least as big as the actual size, bigger is ok */
-    execute_commands_size = (int)(uintptr_t)alloc_executor - (int)(uintptr_t)execute_commands;
-    alloc_executor_size   = (int)(uintptr_t)get_executor   - (int)(uintptr_t)alloc_executor  ;
-    get_executor_size     = (int)(uintptr_t)set_executor   - (int)(uintptr_t)get_executor    ;
-    set_executor_size     = (int)(uintptr_t)print_executor - (int)(uintptr_t)set_executor    ;
-    print_executor_size   = (int)(uintptr_t)print_alloc    - (int)(uintptr_t)print_executor  ;
+    execute_commands_size = (long)(uintptr_t)alloc_executor - (long)(uintptr_t)execute_commands;
+    alloc_executor_size   = (long)(uintptr_t)get_executor   - (long)(uintptr_t)alloc_executor  ;
+    get_executor_size     = (long)(uintptr_t)set_executor   - (long)(uintptr_t)get_executor    ;
+    set_executor_size     = (long)(uintptr_t)print_executor - (long)(uintptr_t)set_executor    ;
+    print_executor_size   = (long)(uintptr_t)print_alloc    - (long)(uintptr_t)print_executor  ;
 }
 
 /* Returns the number of commands not executed */
-int execute_commands(const sandbox_t *box, app_context_t *ctx, int n_cmds) {
+int execute_commands(const sandbox_t *box, app_context_t *ctx, long n_cmds) {
     command_executor_t executor;
     const command_t *cmd;
     int ret;
@@ -42,16 +42,16 @@ int execute_commands(const sandbox_t *box, app_context_t *ctx, int n_cmds) {
 
 int alloc_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cmd) {
     const char *arrname = cmd->args[1];
-    int arrsize = util_strtod(cmd->args[2]);
+    long arrsize = util_strtod(cmd->args[2]);
 
     if(ctx->n_arrays >= MAX_ARRS)
         return -1;
 
-    for(int i = 0; i < ctx->n_arrays; i++)
+    for(long i = 0; i < ctx->n_arrays; i++)
         if(util_strcmp(ctx->arrays[i].name, arrname) == 0)
             return -1;
 
-    int *alloc_base = ctx->allocator((sandbox_t *)box, arrsize * sizeof(int));
+    long *alloc_base = ctx->allocator((sandbox_t *)box, arrsize * sizeof(long));
 
     util_strcpy(ctx->arrays[ctx->n_arrays].name, arrname);
     ctx->arrays[ctx->n_arrays].base = alloc_base;
@@ -62,14 +62,14 @@ int alloc_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cm
 
 int get_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cmd) {
     const char *arrname = cmd->args[1];
-    int idx = util_strtod(cmd->args[2]);
+    long idx = util_strtod(cmd->args[2]);
     const char *varname = cmd->args[3];
 
     if(!arrname || !varname)
         return -1;
 
-    int *arrbase = NULL;
-    for(int i = 0; i < ctx->n_arrays; i++)
+    long *arrbase = NULL;
+    for(long i = 0; i < ctx->n_arrays; i++)
         if(util_strcmp(ctx->arrays[i].name, arrname) == 0) {
             arrbase = ctx->arrays[i].base;
             break;
@@ -77,8 +77,8 @@ int get_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cmd)
     if(!arrbase)
         return -1;
 
-    int *varvalue = NULL;
-    for(int i = 0; i < ctx->n_vars; i++)
+    long *varvalue = NULL;
+    for(long i = 0; i < ctx->n_vars; i++)
         if(util_strcmp(ctx->vars[i].name, varname) == 0) {
             varvalue = &ctx->vars[i].value;
             break;
@@ -100,14 +100,14 @@ int get_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cmd)
 
 int set_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cmd) {
     const char *arrname = cmd->args[1];
-    int idx = util_strtod(cmd->args[2]);
-    int value = util_strtod(cmd->args[3]);
+    long idx = util_strtod(cmd->args[2]);
+    long value = util_strtod(cmd->args[3]);
 
     if(!arrname)
         return -1;
 
-    int *arrbase = NULL;
-    for(int i = 0; i < ctx->n_arrays; i++)
+    long *arrbase = NULL;
+    for(long i = 0; i < ctx->n_arrays; i++)
         if(util_strcmp(ctx->arrays[i].name, arrname) == 0) {
             arrbase = ctx->arrays[i].base;
             break;
@@ -127,8 +127,8 @@ int print_executor(const sandbox_t *box, app_context_t *ctx, const command_t *cm
     if(!varname)
         return -1;
 
-    int *varvalue = NULL;
-    for(int i = 0; i < ctx->n_vars; i++)
+    long *varvalue = NULL;
+    for(long i = 0; i < ctx->n_vars; i++)
         if(util_strcmp(ctx->vars[i].name, varname) == 0) {
             varvalue = &ctx->vars[i].value;
             break;
@@ -147,7 +147,7 @@ void print_alloc(command_t cmd) {
         (uintptr_t)cmd.args[1], 
         (uintptr_t)util_strtod(cmd.args[2])
     };
-    int size = util_snprintf(buf, 256, "allocate array named %s, size %d\n", args);
+    long size = util_snprintf(buf, 256, "allocate array named %s, size %d\n", args);
     prints(buf, size);
 }
 
@@ -158,7 +158,7 @@ void print_get(command_t cmd) {
         (uintptr_t)util_strtod(cmd.args[2]), 
         (uintptr_t)cmd.args[3]
     };
-    int size = util_snprintf(buf, 256, "read array %s index %d into variable %s\n", args);
+    long size = util_snprintf(buf, 256, "read array %s index %d into variable %s\n", args);
     prints(buf, size);
 }
 
@@ -169,7 +169,7 @@ void print_set(command_t cmd) {
         (uintptr_t)util_strtod(cmd.args[2]), 
         (uintptr_t)util_strtod(cmd.args[3])
     };
-    int size = util_snprintf(buf, 256, "set array %s index %d to %d\n", args);
+    long size = util_snprintf(buf, 256, "set array %s index %d to %d\n", args);
     prints(buf, size);
 }
 
@@ -178,6 +178,6 @@ void print_print(command_t cmd) {
     uintptr_t args[] = {
         (uintptr_t)cmd.args[1]
     };
-    int size = util_snprintf(buf, 256, "print variable %s\n", args);
+    long size = util_snprintf(buf, 256, "print variable %s\n", args);
     prints(buf, size);
 }
